Add files via upload
@mpcjanssen Finally I finished the german translation of the help-files. You were right, it took quite some time :-)

The following files I didn't translate, because they are not linked to the help and I am not sure, if they are still in use: defertasks, design, gradle, versions. The changelog I didn't translate neither, doesn't make sense.

I wrote some comments directly into the files and marked them with ???.
Comments or questions to some help files:
addtasks: headers without text, is there missing something? I deleted them for now or did you plan to add something? 
extensions: I suggest to separate the section on the todo.txt-extensions from addtasks and index and put it into a new file. This way the text is only kept in one place. The link to topydo is broken.
index: through the separation of the extensions, it concentrates more on the summary, support-links and Help-Index.
intends: there is no word wrap in Android-Studio, maybe because of the chart?
MYN: Should I update s