Refactor Priority.kt and prio dialog bugfix
Refactor: Kotlin enums give a lot of things for free. We don't need to
duplicate them. Also, there was some unnecessarily general code.

Bugfix: Currently, if you select multiple tasks with Priority A and try
to change their priority, the dialog will be pre-filled with '-' as
their priority. This makes it so 'A' will be pre-filled. If the tasks
have different priorities, the pre-fill will still be '-'.