Make the list wrap datebars items on click, when one of them is too long
This commit will make the task list items change their layout, when one
of the datebars item takes more than one line to display. The changed
layout presents the recurrence item on a new line of its own ; and if
there still are datebars item that takes more than one line, then the
threshold date will go on a new line by itself.

There is a problem as the layout updates only when it would otherwise
be by the UI, that is, when scrolled or touched. I need to find a way
to automatically redraw the list when one multiline datebar item is
detected.